+++ src/components/index.ts
@@ -18,6 +18,6 @@ export { default as Alert } from './alert/alert.svelte';
 export { default as Chip } from './chip/chip.svelte';
 export { default as Card } from './card/card.svelte';
 export { default as Toolbar } from './toolbar/toolbar.svelte';
-export { default as Navbar } from './navbar/navbar.svelte';
-// generic
+export { default as Appbar } from './appbar/appbar.svelte';
 export { default as Spacer } from './spacer/spacer.svelte';
+export { default as Navbar } from './navbar/navbar.svelte';