+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.8]
+
+### Added
+
+- Create **Spacer** component
+- Create **Appbar** component
+- Create **Navbar** component
+
+### Changed
+
+- New logic for utils class devices
+
+### Fix
+
+- Delete duplicate card css on **Toolbar**
+- Size compact **Toolbar** and **Btn** not working
+- Fix size **List** on orientation `horizontal`
+- Fix size ref for **Btn**
+
 ## [0.1.7] - 2025-05-18
 
 ### Added